annealing/agents/requirements.txt
netifaces==0.11.0
PyYAML==6.0.2
pythonping==1.1.4

